requirements.txt
streamlit
google-generativeai
google.generativeai
python-dotenv